refactor: Moved queries into seperate query files.
