+++ src/lib/queries/deleteSession.js
@@ -1,0 +1,9 @@
+export default function getQueryDeleteSession(gql) {
+	return gql`
+		mutation DeleteSessie($sessionId: String!) {
+			deleteSessie(where: { sessieId: $sessionId }) {
+				id
+			}
+		}
+	`;
+}
