+++ src/lib/queries/session.js
@@ -1,0 +1,17 @@
+export default function getQuerySession(gql) {
+	return gql`
+		query GetSessie($sessionId: String!) {
+			sessie(where: { sessieId: $sessionId }) {
+				id
+				sessieId
+				houdbaarTot
+				gebruikerId {
+					id
+					email
+					gebruikersnaam
+					isEmailGeverifieerd
+				}
+			}
+		}
+	`;
+}
