+++ src/lib/queries/updateSession.js
@@ -1,0 +1,9 @@
+export default function getQueryUpdateSession(gql) {
+	return gql`
+		mutation UpdateSessie($sessionId: String!, $expiresAt: Date!) {
+			updateSessie(where: { sessieId: $sessionId }, data: { houdbaarTot: $expiresAt }) {
+				id
+			}
+		}
+	`;
+}
